requirements.txt
# Interfacing with the RPi
RPI.GPIO==0.6.2
